+++ beaker/data_model/dataset.py
@@ -1,5 +1,5 @@
 from datetime import datetime
-from typing import Optional, Tuple
+from typing import Optional, Tuple, Union
 
 from pydantic import validator
 
@@ -59,9 +59,55 @@ class DatasetStorageInfo(BaseModel):
         return v
 
 
-class FileInfo(BaseModel):
+class Digest:
+    SHA256 = "SHA256"
+
+    def __init__(self, digest: Union[str, bytes]):
+        def encode(b: bytes) -> str:
+            import base64
+
+            return f"{self.SHA256} {base64.standard_b64encode(b).decode()}"
+
+        encoded: str
+        if isinstance(digest, bytes):
+            encoded = encode(digest)
+        else:
+            encoded = digest
+
+        self._encoded = encoded
+
+    def __eq__(self, other) -> bool:
+        if isinstance(other, Digest):
+            return self.decode() == other.decode()
+        elif isinstance(other, str):
+            return self == Digest(other)
+        elif isinstance(other, bytes):
+            return self.decode() == other
+        else:
+            return False
+
+    def __ne__(self, other) -> bool:
+        return not self == other
+
+    def __str__(self) -> str:
+        return self._encoded
+
+    def __repr__(self) -> str:
+        return self._encoded
+
+    def decode(self) -> bytes:
+        """
+        Decode a digest into its raw bytes form.
+        """
+        import base64
+
+        encoded = self._encoded.split(" ", 1)[-1]
+        return base64.standard_b64decode(encoded)
+
+
+class FileInfo(BaseModel, arbitrary_types_allowed=True):
     path: str
-    digest: str
+    digest: Digest
     updated: datetime
 
     size: Optional[int] = None
@@ -73,6 +119,13 @@ class FileInfo(BaseModel):
     """
     A URL that can be used to directly download the file.
     """
+
+    @validator("digest", pre=True)
+    def _validate_digest(cls, v: Union[str, Digest]) -> Digest:
+        if isinstance(v, Digest):
+            return v
+        else:
+            return Digest(v)
 
 
 class DatasetManifest(BaseModel):