+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Added `Digest` type and changed type of `FileInfo.digest` from `str` to `Digest`.
+
 ## [v0.15.2](https://github.com/allenai/beaker-py/releases/tag/v0.15.2) - 2022-05-09
 
 ### Added